Split views tests, update facade default to webcomponents-1.0.5
Split monolithic page_registry_test.lua into focused test files:
- page_registry_test.lua (page discovery, detail, proxy, template, access, kind)
- component_registry_test.lua (find_all, get, entry_point, sorting)
- resource_registry_test.lua (find_all, group_by_type, collect_for_page, globals)
- env_loader_test.lua (build_env_context, validate_priority, get_recommended_priority)

Update facade fe_facade_url default from 1.0.4 to 1.0.5.
