+++ src/facade/_index.yaml
@@ -33,7 +33,7 @@ entries:
     targets:
       - entry: wippy.facade:fe_facade_url
         path: .default
-    default: https://web-host.wippy.ai/webcomponents-1.0.4
+    default: https://web-host.wippy.ai/webcomponents-1.0.5
 
   - name: fe_entry_path
     kind: ns.requirement
